requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.2.3
h11==0.14.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
pycparser==2.21
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
Werkzeug==2.2.3
wsproto==1.2.0
